+++ app/app.rb
@@ -5,44 +5,5 @@ require 'sinatra/flash'
 
 require_relative 'data_mapper_setup'
 
-class Chitter < Sinatra::Base
-
-  register Sinatra::Flash
-
-  enable :sessions
-  set :session_secret, 'super secret'
-
-  get '/' do
-    erb :'index'
-  end
-
-  get '/users/sign_up' do
-    @user = User.new
-    erb :'users/sign_up'
-  end
-
-  post '/users' do
-    @user = User.new(first_name: params[:first_name],
-                last_name: params[:last_name],
-                email: params[:email],
-                password: params[:password],
-                password_confirmation: params[:password_confirmation],
-                user_name: params[:user_name])
-    if @user.save
-      session[:user_id] = @user.id
-      redirect('/')
-    else
-      flash.now[:errors] = @user.errors.full_messages
-      erb :'users/sign_up'
-    end
-  end
-
-  helpers do
-    def current_user
-      @current_user ||= User.get(session[:user_id])
-    end
-  end
-
-  # start the server if ruby file executed directly
-  run! if app_file == $0
-end
+require_relative 'server'
+require_relative 'controllers/users'
