+++ app/controllers/users.rb
@@ -1,0 +1,24 @@
+class Chitter < Sinatra::Base
+
+  get '/users/sign_up' do
+    @user = User.new
+    erb :'users/sign_up'
+  end
+
+  post '/users' do
+    @user = User.new(first_name: params[:first_name],
+                last_name: params[:last_name],
+                email: params[:email],
+                password: params[:password],
+                password_confirmation: params[:password_confirmation],
+                user_name: params[:user_name])
+    if @user.save
+      session[:user_id] = @user.id
+      redirect('/')
+    else
+      flash.now[:errors] = @user.errors.full_messages
+      erb :'users/sign_up'
+    end
+  end
+
+end
